Removing particles, feedback improvements, start of gui.

+++ client/src/main/scala/json/BaseSerializers.scala
@@ -16,7 +16,7 @@ object BaseSerializers {
   }
 
   implicit val intOptionWriter = Writer[Option[Int]] {
-    case Some(i) => Js.Num(i)
+    case Some(i) => Js.Num(i.toDouble)
     case None => Js.Null
   }
   implicit val intOptionReader = Reader[Option[Int]] {
